Refactoring python tracing example
