+++ .github/workflows/unit_tests.yml
@@ -48,7 +48,7 @@ jobs:
       run: ./scripts/install_pmutools.sh
 
     - name: Get extra python packages
-      run: pip install -r ./function-images/tests/tracing/python/client-server/requirements.txt
+      run: pip install -r ./function-images/tests/tracing/python/integ-tests/client-server/requirements.txt
 
     - name: Build
       run: go build -race -v -a ./...
@@ -94,4 +94,23 @@ jobs:
     - name: Cleaning
       if: ${{ always() }}
       run: ./scripts/clean_fcctr.sh
-      
+
+  bench-python-tracing:
+    name: Python Tracing tests (docker-compose)
+    runs-on: ubuntu-18.04
+    defaults:
+      run:
+        working-directory: ./function-images/tests/tracing/python/integ-tests/client-server
+    strategy:
+      fail-fast: false
+    steps:
+    - uses: actions/checkout@v2
+    
+    - name: build local images
+      run: make all-image
+    - name: Python client with docker-compose
+      run: |
+        docker-compose up --exit-code-from client
+    - name: Golang client with docker-compose
+      run: |
+        docker-compose --file ./docker-compose-goclient.yml up --exit-code-from client
