+++ function-images/tests/tracing/python/integ-tests/client-server/docker-compose-goclient.yml
