+++ function-images/tests/tracing/python/integ-tests/client-server/docker-compose.yml
