+++ function-images/tests/tracing/python/integ-tests/client-server/goclient/go.mod
