+++ function-images/tests/tracing/python/integ-tests/client-server/helloworld_pb2_grpc.py
