+++ function-images/tests/tracing/python/integ-tests/client-server/requirements.txt
@@ -1,5 +1,5 @@
 futures
-grpcio
+grpcio==1.38.1
 opentelemetry-api==1.3.0
 opentelemetry-exporter-zipkin==1.3.0
 opentelemetry-exporter-zipkin-json==1.3.0
